Fix preload query for through associations with custom scope block
Using a through association with `includes` could raise an exception if
the association had a custom scope block. This was because the preload
query for the through association would not join in the source
association's table, but the conditions in the where clause could depend
on that table.

This changes `ActiveRecord::Associations::Preload#through_scope` to
also spawn references for the reflection_scope's source if there was a
where clause on the reflection_scope. This makes the query join in the
source association's table so any conditions in the where clause will
have the table to reference.