Add query caller config to AR Core & LogSubscriber
The ActiveRecord::Core config defaults to false because it’s a new
behavior that changes the log output quit a bit by displaying the call
site of each database query.

The ActiveRecord::LogSubscriber config is also falsey by default but it
is superseded by the ActiveRecord::Core configuration.

It is however enabled by default in the development environment
configuration, which is admittedly debatable.

Coupling with ActiveSupport::BacktraceCleaner was removed and the
LogSubscriber mock was modified to allow the debug method to be called.