Update system test runner docs
It wasn't clear that system tests don't run with the rest of the test
suite and are part of a separate command.

This documents the `test:system` command as well as update the Rails
runner help documentation to make it clearer that system tests are run
via their own command by default.