`changes` should not be a HWIA
To make a long story as short as possible, if an attribute is an Array
of Hashes (I believe just a normal Hash as the value should also cause
this behavior), it will be converted to a HWIA through this method. This
is because when you pass the above value to `HWIA[]=`, it will convert
all nested Hashes to be HWIA's.

The easiest way to solve this is to have
`changes` not be a HWIA anymore, but that breaks the "indifferent
access" part of the API contract and the tests that watch over that
behavior. Maybe we need to create a mini-class for only indifferent
access, and not worry about setting values, since that's not intended
behavior here?

This commit fixes a regression between Rails 5.0 and 5.1, which was
introduced by way of 16ae3db5a5c6a08383b974ae6c96faac5b4a3c81. That
commit technically didn't introduce the truly regression causing behavior --
it was the "light switch" commit that directed users toward the bad
behavior.