Ignore errors while loading db/structure into postgresql
This will allow the loading of db/structure.sql files with non-fatal
errors (like "set row security" or adding schema COMMENTs) to complete
without failure, similar to the behavior under rails < 5.

To use this, export IGNORE_PG_LOAD_ERRORS=1 when loading your
db/structure.sql, thusly:

    bundle exec rails IGNORE_PG_LOAD_ERRORS=1 db:structure:load

This is necessary because many rails developers work with postgres
connections that have superuser access, and the process of dumping
`structure.sql` will reflect that privilege level. This typically
happens on osx machines. This makes it hard to interoperate with
developers that use postgres connections that aren't configured as the
superuser - for instance, most default linux installs and most CI providers.

This isn't a perfect solution, but it allows us to at least function
similar to pre rails 5, where we'd do a "best effort" loading of the
structure. This generally worked OK.