Drop DebugExceptions out of the default middleware stack
On production, the `ActionDispatch::DebugExceptions` middleware just
delegates the error to the `ShowExceptions` middleware. In fact, it
kinda does the job of something we already have pluggable inside of
`ShowExceptions`. It's `exception_app`.

With a bit of juggling, we are able to use the `DebugExceptions`
middleware as an `exception_app` and not a Rack middleware. This is
nice, as it will free a chain of the middleware stack. However, it will
force a lot of plugins to change, because many of the error handling
gems do use `ActionDispatch::DebugExceptions` as a reference to where
to insert their own middlewares.

This is still a WIP and I will describe the idea in a bit more detail
soon.