Document usage of QueryMethods#references with QueryMethods#or
QueryMethods#or prohibits to connect relations with different references.
However, it should be still valid if the author of the query is aware that
references in the relations connected by #or, requires to add JOIN in the
whole connected query.

QueryMethods#references is an appropriate method for developers to show
they know such consequences.