Resolve #29202 Rails or rails generated project can't bundle
coffee-rails gem is not compatible with rails 5.2 and it stops
any apps made using rails 5.2 on edge or dev from bundling.
The coffee-rails repo has fixed this but not cut a new release
since rails 5.2 is not in main ciruclation. It is reasonable to
avoid this problem in the future the edge and dev versions of
apps generated uing rails 5.2 should use the edge version of
the coffee-rails gem