Add notice about migration versioning in 5.1
When I started using 5.1 it was very surprising for me that I got an
error when I tried to run my migrations. And there is not that many
information about what happened. There is some blog post somewhere but
nothing on the official channel. I guess it would be useful to have it
here.