Improve Fixture Caching for Same Table Fixtures
Problem

``` ruby
class A < ActiveRecord::TestCase
  fixtures :parrots, :cars

  def test_fixtures_loaded
    # Dusty may or may not be present, depending on whether some other
    # test has loaded the :live_parrots fixture.
    assert_equal [nil, "Curious George", "King Louis", "frederick", polly"], Parrot.order(:name).pluck(:name) - ["Dusty Bluebird"]
  end
end

class B < ActiveRecord::TestCase
  fixtures :parrots, :live_parrots, :cars

  def test_fixtures_loaded
    assert_equal [nil, "Curious George", "Dusty Bluebird", "King Louis", "frederick", "polly"], Parrot.order(:name).pluck(:name)
  end
end
```

If Class B runs after class A, class B's test will fail. Because
:parrots is already cached, it skips the fixture, but when it gets to
:live_parrots, it drops the table and loads up the :live_parrots
fixtures.

Solution

Keep a map of fixtures by table. If when it comes time to create
fixtures and the fixtures for the table match what is cached, then we
can skip creating them in the database. If the fixtures are short of
what we have cached, then the database needs to be cleaned to remove old
records.

See the conversation on https://github.com/rails/rails/pull/29298 for
more information.