Drop IRB out of the web-console Gemfile comment
We don't provide exactly the same experience as IRB does, so let's not
advertize it like that. We can say that it's an interactive console,
without further references. I have also followed byebug's comment and
changed the calling `console` part. Hopefully, this can hint that the
invocation is not view specific and you can use it like the debugger
statement.

[ci skip]