Fix id resolution bug by changing id_in_database method name to id_in_db
Closes #29436

This was happening because Active Model creates methods with `_in_database` suffix
for all attributes of the model. While saving the model, in persistence module
Active Record tries to call `id_in_database` method which is defined in `PrimaryKey`
module to get the primary_key value of the object but this call ends in a method which
is defined by Active Model `AttributeMethods` for id attribute of the object.
Renaming `id_in_database` method fixes the problem.