Fix a problem when re-defining has_many :through associations
This is an attempt to fix this issue:
https://github.com/rails/rails/issues/29123

When calling a ‘has_many :through’ association we
check if the source association is defined after
the has_many :through. If it is, we throw a
HasManyThroughOrderError.
(See https://github.com/rails/rails/commit/f8ab3ae18fbb5c4a4c9563296d0e7c528e754c42)

However, this error was being erroneously thrown
in cases when the has_many :through overrides
another association of the same name.