Fix system test generator option to be consistent
We use `skip-system-test` everywhere else so it doesn't make sense to
use `system_tests = nil` in the generator config. Otherwise we're
passing conflicting options around (`system_tests` by default
`:test_unit` otherwise `nil` vs `skip_system_test` default is false and
otherwise pass true). It gets confusing.

Because of this change we then need to apply `skip_system_test` to the
`system_test` hook. In the PR to generate a more complete system test
when generating a scaffold this will be changed to be generated with
`create_test_files` but for now this works to unblock that PR.