Apply record state based on parent transaction state
Let's say you have a nested transaction and both records are saved.
Before the outer transaction closes, a rollback is performed. Previously
the record in the outer transaction would get marked as not persisted
but the inner transaction would get persisted.

```ruby
Post.transaction do
  post_one.save # will get rolled back

  Post.transaction(requires_new: true) do
    post_two.save # incorrectly remains marked as persisted
  end

  raise ActiveRecord::Rollback
end
```

To fix this the PR changes transaction handling to have the child
transaction ask the parent how the records should be marked. When
there are child transactions, it will always be a SavpointTransaction
because the stack isn't empty. From there we pass the parent_transaction
to the child SavepointTransaction where we add the children to the parent
so the parent can mark the inner transaction as rolledback and thus mark
the record as not persisted.

`update_attributes_from_transaction_state` uses the `completed?` check to
correctly mark all the transactions as rolledback and the inner record as
not persisted.

```ruby
Post.transaction do
  post_one.save # will get rolled back

  Post.transaction(requires_new: true) do
    post_two.save # with new behavior, correctly marked as not persisted
    on rollback
  end

  raise ActiveRecord::Rollback
end
```

Fixes #29320