tld_length of setting a cookie consistent with other tld_length
With the `tld_length` passed as an argument of setting a cookie, the
value refers to the length of the tld plus the domain itself. So to
share cookies across a.rubyonrails.co.uk and b.rubyonrails.co.uk, you'd
set the tld_length to 3.

With `config.action_dispatch.tld_length`, the value refers only to the
length of the tld (non-domain) portion of the domain. So for
www.rubyonrails.co.uk, tld_length should be set to 2.

As action dispatch's `tld_length` is more logical, and most likely more
widely used, this changes the `tld_length` when setting a cookie to
carry the same meaning.

As this is a non-backwards compatible change, it introduces a new config
value so that existing applications can work as before.