Eager loading with scoping should behave consistently
Currently `Post.find(1).comments` and with `.eager_load(:comments)`
behave consistently in the scoping. But with `.includes(:comments)` or
`.preload(:comments)` ignores the scoping. This means that we cannot
rewrite from `Post.find(1).comments` or with `.eager_load(:comments)` to
with `.includes(:comments)` or `.preload(:comments)` if affected in the
scoping. It should be behaved consistently.