Do not execute `secrets_test.rb` in parallel
In encrypted secrets, the tmp file is used as a fixed  file
(`secrets.yml.enc` under the tmp directory).
And that tmp file will be removed after process.

Therefore, if test in parallel, there is a possibility to write to the same
tmp file, or to remove the tmp file being used by another process yet.
This occasionally caused the test to be an error.

For this reason, modified not to run in parallel.