Catch postgres connection errors when trying to dealloc the statement pool
connection_active? will sometimes return true when the connection is actually dead/disconnected (see #3392 for a discussion of why this is).  When this happens, a query is run on the dead connection which causes various postgres connection errors to be raised.  This fix catches any such errors and ignores them.

Closes #29760