Post.joins(:users) should not be affected by `User.current_scope`
This change was introduced by #18109. The intent of that change was to
specifically apply `unscoped`, not to allow all changes to
`current_scope` to affect the join. I'm not yet sure if I want to keep
that change in behavior on 5.0 and later (it's been around for 2
releases now). However, we *certainly* cannot have the change in
behavior affect 4.2.

Fixes #29338.