Fix build failures on PG
Honestly I don't think the tests that are fixed by this change should
have been merged. Passing a range or an array to `where` has a special
meaning. We need to solve the problem more concretely without overriding
the behavior that is present for *every* other type.

However, the damage has been done. These changes were in 5.1, so we need
a deprecation cycle to remove it.