Allow bin/rails test task to take absolute paths as arguments
Solves #29923

This regression was caused due to a wrong regex to filter out
paths, introduced in commit 796a1cf0e

The regex was /^\w+\// which did not accept paths with a leading
slash and hence all absolute paths were filtered out.

This change introduces a change in regex which allows for a leading
slash and acts on the matched term accordingly.

While cascading through the case block, the paths are checked for
line number specification, existence of a directory at that path
and if none of those match, then it is considered to be a path to the
file. The regex matchers specified are filtered out via the call
to `Array#compact` since they do not match any of these conditions.