Remove unused Signal trap for INT
This removes unused/unreferenced trapping of `INT` Signal and exiting
rails server. This is unused because `Rails::Server` inherits from
`::Rack::Server` and `Rails::Serve#start` invokes `super` which overrides the
trapping of `INT` to what is already declared in `::Rack::Server#start`.

Hence keeping this here leads to confusion.