Revert "Fix path encoding tests if Ruby version is 1.9.3"
This reverts commit 31dfc4f99211b9e3a8675b2983389f57b037e5f8.

A bad combination of commits required this to be backported. The
original change to the tests here shouldn't have been backported. The
change to Rails master was technically incorrect.

What happened was `@tenderlove` and I made a change to Rack and then
wrote this commit to accompany it: #29062. This test was changed after
that and backported to Rails 4.2. It for some reason wasn't backported
to 5.0 or 5.1.

Then on release day we realize the build was failing for Ruby 2.0+ on
Rails 4.2 so I ended up backporting #29062 when it shouldn't have been.
The correct behavior was to not backport #29062 or #29095 to Rails 4.2.

This revert fixes that.