Revert "Revert "Merge pull request #29062 from eileencodes/force-encoding-to-original-string-encoding""
This reverts commit aa733278579a45f750bbabd10bf8ec2863215762.

A bad combination of commits required this to be backported. The
original change to the tests here shouldn't have been backported. The
change to Rails master was technically incorrect.

What happened was `@tenderlove` and I made a change to Rack and then
wrote this commit to accompany it: #29062. This test was changed after
that and backported to Rails 4.2. It for some reason wasn't backported
to 5.0 or 5.1.

Then on release day we realize the build was failing for Ruby 2.0+ on
Rails 4.2 so I ended up backporting #29062 when it shouldn't have been.
The correct behavior was to not backport #29062 or #29095 to Rails 4.2.

This revert fixes that.