Actually revert changes to encoding that were backported
I don't know what happened when I reverted this but the revert in the
previous commit didn't revert everything which caused Ruby 1.9.3 to not
be correct and the test to be incorrect.

This restores the previous behavior that was expected before this test
changed what it though the expectations of Rack were. Requests that have
incorrect encoding should return a `:bad_request` and not `:ok` or
`:redirect` or `:not_found`.