Buffer received messages in client socket
AC clients do not receive update about state of a subscription.

I originally assumed this might be related to #24874 and applied the WIP patch by
@maclover7. This did not solve the problem. I've investigated some more and I
think the the crux of the problem is:

The cables client socket is open and receives a subscription message, which is
discarded since the socket is not in the desired ready state at that point.

This leaves both server and client in a stalled state. The client is waiting for
a subscription update, the server waiting for a subscription message.

While debugging this issue I've documented my findings in this
[project](https://github.com/wpp/action-cable-debugging), which contains a Rails
app & the required scripts to reproduce the problem. Please see the README for
more information.