Respect layout option when rendering collection
When doing a partial render for a collection there is a layout option
available. It is used to wrap each rendered partial. This works as
expected in all cases (that I've found) except when the collection is
heterogeneous - or at least when the objects in the collection result in
a set of non-unique `to_partial_path` values.

A change back in 228f9910389cad7fe0dd7f2bd010fe654f794b37 caused this
inconsistency. The comment there says the layout option is ignored for
the shortcut `render @collection`, but makes no mention of the longer
form. There has been much churn in this area in the years since that
change, and it seems we ought to respect the option as it's consistent
with the existing documentation, and other ways of calling this partial
render.