Change all OSX/OS X/Mac references to macOS
Since Apple renamed the operating system, it makes sense for us to bring
our documentation and library code in line with that rename.

NOTE: I was going to directly commit the documentation changes to
master, and then PR the library code changes separately, but I figured
it was better for this change to happen together/all at once.