Decouple the merge/normalization and conditional cache control logic
The prior logic was trying to do too many things at once. For all responses,
we want to perform two distinct steps:

* Merge/normalize the `Cache-Control` values found in HTTP headers and those
  found in the `@cache_control` hash.
* Conditionally set a default `Cache-Control` header value when we have an ETag

This change separates these concerns since the merge/normalize step should
occur for all responses, but the second should only occur when we have already
set an ETag/last modified value. Normally ETag middleware will set a default
`Cache-Control`, but only if an existing ETag is not already set. So, in the
cases where an ETag is set, we need to set the default `Cache-Control` value
ourselves.