Appropriately request to unset and set cookie for different sub-domains
Currently if we were to delete a cookie and set a same cookie but with a
different sub domain, only the new cookie gets set. The cookie that is marked
for removal is still persisted. For example:

```ruby
cookies.delete(:foo, domain: "bar.com")

cookies[:foo] = { value: "bar", expires: 1.year.from_now, domain: "xyz.bar.com" }
```

In the above scenario, single `Set-Cookie` header is sent for `xyz.bar.com`.
Which leaves the earlier cookie persisted in browser, because technically, those
are two different domains.

This may not be a very common use case, but is helpful when working with
legacy cookie and introducing new sub domains.

Hence, this PR. Where, in such a scenario we now send two `Set-Cookie` headers.
One to remove/unset the old cookie and the other to set the new cookie
value on new domain.

I have added new tests and extended an existing test to make sure existing
behaviors are persisted.