Remove Active Storage duplicated classes from the API site
Since cb5af0d7, some classes that are under Active Storage are now
part of the API site.

However, these classes aren't nested under a definition of the
`ActiveStorage` module but rather name-spaced under it like
`ActiveStorage::Foo`.

Thus, these classes are present both under the ActiveStorage label
and at the root of the site's sidebar so we have to strip out
duplicates.

[ci skip]