Allow for partial durations via `/` operator
Previously durations could be divided and would return partial durations,
but in Rails 5.1 dividing a duration such as

    1.day / 2

would return a duration of 0 instead of 0.5 days. This change restores
that functionality by forcing the divisor to a float, which allows for
the partial duration to be returned. Also updated Duration::Scalar to
behave the same way. Fixes #30546.