Add `datalist` option to text_field helpers
This commit introduces a `datalist` option to the form helper `text_field` and form tag helper `text_field_tag`, allowing the
caller to pass a list of options used to construct a HTML datalist
element, and associate it with the input.

With several input types delegating to text_field helpers, the 
`datalist` option also works for range_field, number_field, color_field and the time & date related helpers.