Remove `include FileUtils` from bin templates
## Summary

When creating a new Rails application with `rails new`, a top-level
`include` is used in two files `bin/setup` and `bin/update`.

This PR makes `FileUtils` an explicit receiver instead of a top-level
`include`. This change removes the top-level `include` affecting
`Object` from the templates of a Rails application created with `rails new`.

By writing `include FileUtils` at the top level, it can call abbreviated
methods like `chdir`, `cp` and others, but here it seems that there is
not much merit. These are templates for short scripts of processes, but
I think that it is preferable not to affect `Object`.

## Other Information

I found it by investigating a Rails applications when making
https://github.com/bbatsov/rubocop/pull/4840.