WIP validate a field only if that field has changed
This commit is a proof of concept for running validations on a field only if that field has changed.
First of all targeting those validations that require a database query to perform. E.g. association presence and uniqueness validations.

As a side effect this addresses problems like @schneems had with [codetriage](https://github.com/codetriage/codetriage/pull/573#issue-239516618) and which is described in his blog post named [How I Reduced my DB Server Load by 80%](https://schneems.com/2017/07/18/how-i-reduced-my-db-server-load-by-80/).

In my view it makes sense to allow users to choose between the default "always-validate" behaviour and this kind of the "validate-changed" behaviour. An option and a config option similar to [`belongs_to`](https://github.com/rails/rails/pull/18937) would be nice to have.

There is an obvious speed improvement, especially for large tables.
A downside is a rare case when you need to validate a field even if there is no change (any examples?).

Thoughts?