Include reference key type exception for SQLite
Before, all primary keys had type integer by default and the
documentation reflected that. After #26266 MySQL and PostgreSQL
started using +:bigint+ as default type, while SQLite continued
using +:integer+ as big integer is not supported.

At #31007 documentation was updated to +:bigint+, but this
was not accurate because SQLite does not support it. This
commit reflects this exception.