Add new config for skip_callback
After Rails 5 if we use `skip_XXX_callback :foo` before foo is being
setted as a callback, we'll get argument error. The only way to avoid
this is to pass `raise: false` to every `skip_XXX_callback` call, which
can be very annoying.

So I add a new config for ActiveSupport::Callback, called
`skip_callback_raise_exception_if_not_defined`. It's purpose is to let
users config if they want the exception get raised globally instead of
passing options in every skip callback call. And the default value is
true so the current behavior will remain the same.