Lock faye-websocket on 5-0-stable branch
faye-websocket 0.10.7 was breaking the tests on 5-0-stable. Since faye
is removed in 5.1 and 5.0 is going to only be receiving security patches
soon I think locking the gem is a fair compromise.

If you're using faye and action cable in 5.0 and experience an issue
with the updated faye-websocket open an issue on GitHub.