Previously collection models used to get persisted even when there's an explicit ActiveRecord::Rollback in a callback and change the count of the collection obect without any regard to the callback specification.
This used to happen because ActiveRecord::Rollback used
to be silently swallowed and the rest of the pipeline triggered
commit from it's previously assigned data values.

For example:

@model.collection.create would yield 1 even if there's ActiveRecord::Rollback
in after_create callback of Collection model.

This PR fixes the problem and makes the count consistent
if there's a rollback on a collection object.

Fixes #28836 and #28536