Fix root engine URL helpers not generating with a subdirectory
This references the following RAILS issue that has a sample application and
further expands upon the content of this pull request:
https://github.com/rails/rails/issues/31476

This fix is a bit hacky as the way subdirectories are handled needs to be
cleaned up. Essentially the issue boils down to the following:

* In order to support multiple engine mount points, a function is run to strip
away the last "/" of an engine's path. Without a subdirectory, this would be
the engine name (ie. "/myengine") if a value existed there.

* However, with a subdirectory configured there, that last value isn't always
the engine name. Rather, it can be the subdirectory itself if the engine is
mounted to the root path. An example of this would be something like
"/subdirectory" that then causes that method to incorrectly remove
the "subdirectory" value. (NOTE: This did not effect non-root engines as the
logic would still work for them, ie: "/subdirectory/myengine" would correctly
remove the "myengine" mount point).

* This fix checks the setting "relative_url_root" and prevents that value from
being stripped out if the engine reports itself to be mounted at "/". However,
the only way this value will work is if ENV["RAILS_RELATIVE_URL_ROOT"] is set
rather than it being configured manually in the root Rails application config.
This is due to the engine using default values only for its mounting.

* I added a Unit Test based on one that already existed for an engine that wasn't
mounted at root. This test will fail with the old code but pass correctly with
this change.

Thanks!