Introduce `ActiveRecord::Relation#one`
`#one` asserts that the relation contains exactly one record, and
returns it. If there are no records or more than one, an
`ActiveRecord::NoRecordFound` or `ActiveRecord::MultipleRecordsFound`
exception is raised, respectively.

This is a concept borrowed from other ORM tools. Is useful when a
data schema has a "many" relationship, but only a single record is
expected in some situation.

Due to potential confusion with `ActiveRecord::Relation#one?`, I
also suggest the name `#single`.