Remove ActiveModel::Attributes ???
I suspect I have missed something here, but #31832
got me poking around in `ActiveModel::Attributes`, and
I couldn't see where it was being used, except in
a couple tests.

If I am correct that this undocumented
class is not being used anywhere in Rails, we should either
remove it, or document it so people can use it. It does have
some duplication with ActiveRecord::Attributes, so maybe
we could clean that up as well.

If I am NOT correct about any of this, I will be happy to learn why!