Do not check `ActiveStorage::Engine` when updating to 5.2
When upgrading from 5.1 to 5.2, if the application did not require
`rails/all`, `ActiveStorage` not be defined.
In that case, Acttive Storage content is not generated regardless of whether
the user wants to use or not.

As for the new component, I think it is preferable to check whether to
use it for the user, but since it is not yet able to deal with it, I
think that it is good to forcibly generate the content.