Allow env-specific credentials by config
In many cases, credentials is only necessary for production.
However, if use a service like payment, want to change the api key in
staging and production.
By changing the behavior by config, I think that credentials can be used
smoothly even in such a case.

Also, if application using the secret encrypted with 5.1, this will make
it possible to upgrade smoothly.