Use authenticated cookie options to disable embedding of the expiration in the cookies
We were generating the cookie value with the expiry information embedded what
makes the value impossible to be read by a Rails 5.1 application.

This option is only useful if you need to share your cookies between
a Rails 5.1 and a Rails 5.2 application, or if you are still validating
the deploy and wants to eventually rollback.