Fix counter_cache decremented twice
This commit fixes the bug reported in #31493 and #32079.

It includes the following test that would fail without the patch as
follows:

Failure:
BelongsToAssociationsTest#test_decrement_counter_cache_once
[rails/activerecord/test/cases/associations/belongs_to_associations_test.rb:778]:
Expected: -1
  Actual: -2

The fix was proposed in #31494 by "gm_zhou <gm_zhou@ctrip.com>".
The test was added by the author of this commit.

[Guangming Zhou, Malik Olivier Boussejra]