belongs_to_associations_test.rb: Simplify test case
Test case was using changing the a dog's nested attributes using
`accepts_nested_attributes_for`. However, this is not required to
reproduce the bug fixed by 6107b37.

Indeed, the bug can be reproduced simply by having a model with two different
`counter_cache`.

As we do not change the definitions of dogs or dog_lovers, the overall impact
of the PR over the project is reduced by this commit.