Don't mutate parameters in test request helpers
Mutating the value passed as the `params` keyword argument when making
a request in a controller test is surprising behaviour. We can avoid it
by calling `symbolize_keys` first, which will make a copy of the hash.

This matches the behaviour on 5.1 and later, which was introduced in
98b8309569a326910a723f521911e54994b112fb. On 4.2 and earlier, passing
`format` separately wasn't supported, so this fix is only needed on 5.0.